pocitani_minci/backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioquic==1.2.0
aiosignal==1.3.2
alembic==1.13.2
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.2.0
bcrypt==4.1.3
blinker==1.8.2
boto3==1.34.144
botocore==1.34.144
Brotli==1.1.0
build==1.2.1
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.8
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
Deprecated==1.2.14
distlib==0.3.8
ecdsa==0.19.1
Faker==26.0.0
fastapi==0.115.12
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.5.0
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.0.1
idna==3.7
importlib_resources==6.4.0
isodate==0.6.1
itsdangerous==2.2.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
kaitaistruct==0.10
kiwisolver==1.4.5
langchain==0.3.13
langchain-core==0.3.28
langchain-text-splitters==0.3.4
langsmith==0.2.6
ldap3==2.9.1
limits==3.13.0
lxml==5.2.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mitmproxy==11.0.2
mitmproxy-windows==0.10.7
mitmproxy_rs==0.10.7
more-itertools==10.3.0
msal==1.29.0
msal-extensions==1.2.0
msgpack==1.1.0
multidict==6.1.0
numpy==2.0.1
opencv-python==4.11.0.86
ordered-set==4.1.0
orjson==3.10.12
packaging==24.1
passlib==1.7.4
pillow==10.4.0
pipenv==2024.0.1
platformdirs==4.2.2
portalocker==2.10.1
premailer==3.10.0
propcache==0.2.1
publicsuffix2==2.20191221
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydivert==2.1.0
Pygments==2.18.0
uvicorn==0.34.3
waitress==3.0.0
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
yagmail==0.15.293
yarl==1.18.3
zstandard==0.23.0
python-multipart==0.0.6
